Fix ESLint issues
Implements the comprehensive ESLint fixes plan, addressing type safety, unused variables/imports, and Fast Refresh issues. This includes resolving `any` types in `backgroundSync.ts` and WorkOrder components, fixing `no-case-declarations`, removing unused code, and extracting context/hooks to separate files.
