+++ src/test/utils/renderUtils.tsx
@@ -1,43 +1,6 @@
-import React, { ReactElement } from 'react';
+import { ReactElement } from 'react';
 import { render, RenderOptions } from '@testing-library/react';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { MemoryRouter } from 'react-router-dom';
-import { 
-  MockAuthProvider, 
-  MockSessionProvider, 
-  MockUserProvider, 
-  MockSimpleOrganizationProvider,
-  MockSessionProvider2 
-} from './mock-providers';
-
-// Test providers wrapper
-const TestProviders = ({ children }: { children: React.ReactNode }) => {
-  const queryClient = new QueryClient({
-    defaultOptions: {
-      queries: {
-        retry: false,
-      },
-    },
-  });
-
-  return (
-    <MemoryRouter initialEntries={['/']}>
-      <QueryClientProvider client={queryClient}>
-        <MockAuthProvider>
-          <MockSessionProvider>
-            <MockSessionProvider2>
-              <MockUserProvider>
-                <MockSimpleOrganizationProvider>
-                  {children}
-                </MockSimpleOrganizationProvider>
-              </MockUserProvider>
-            </MockSessionProvider2>
-          </MockSessionProvider>
-        </MockAuthProvider>
-      </QueryClientProvider>
-    </MemoryRouter>
-  );
-};
+import { TestProviders } from './TestProviders';
 
 export const customRender = (
   ui: ReactElement,
